LayoutTests/svg/custom/svg-parse-overflow-1.html

<html>
<body>
<script>
document.body.offsetTop;
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<svg>
<foreignObject y="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000" transform="scale(5) skewX(5)">
</svg>
</body>
