LayoutTests/svg/custom/svg-parse-overflow-1-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute y="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
